fix(review): restore arch-compliant structure; narrow exception; fix gate ordering
REQ-IMP-003 restricts tools_*.py from importing autoskillit.recipe and
autoskillit.execution directly, so _find_recipe and _prime_quota_cache must
remain in helpers.py as the architecture-approved bridge.

Changes:
- Restore _find_recipe to helpers.py with a docstring explaining the arch constraint
  (finding 1/3 partially addressed — proxy is justified by REQ-IMP-003)
- Restore _prime_quota_cache to helpers.py with narrowed exception catch:
  bare Exception -> (OSError, ValueError, RuntimeError) to stop swallowing
  programming errors like AttributeError/ImportError (finding 2)
- open_kitchen: call _open_kitchen_handler() before ctx.enable_components() so
  gate state and gate file are established before tools become visible (finding 4)
- close_kitchen: call _close_kitchen_handler() before ctx.reset_visibility() so
  gate cleanup always runs even if reset_visibility raises (finding 5)
